src/saturnin/core/fb_log_parse/pyproject.toml
[build-system]
requires = ["setuptools >= 65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saturnin.firebird.log.parser"
version = "0.2.1"
description = "Firebird log parser microservice"
readme = "README.rst"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [{ name = "Pavel Císař", email = "pcisar@users.sourceforge.net"}]
keywords = ["Firebird", "server", "log", "parser", "micro"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Logging",
    "Topic :: Database"
    ]
dependencies = [
    "saturnin>=0.7.2",
    "firebird-lib>=1.2.2"
    ]

[project.urls]
Home = "https://github.com/FirebirdSQL/saturnin"
Documentation = "https://saturnin-core.rtfd.io"
"Bug Reports" = "https://github.com/FirebirdSQL/saturnin-core/issues"
Funding = "https://www.firebirdsql.org/en/donate/"
Source = "https://github.com/FirebirdSQL/saturnin-core"

[tool.setuptools]
zip-safe = true

[tool.saturnin.metadata]
component-type = "service"
uid = "35fcea0e-c0fd-55a4-9641-12ed0cf85d9b"
descriptor = "api:SERVICE_DESCRIPTOR"
